+++ argos/__init__.py
@@ -2,6 +2,6 @@
 Argos Index - Sistema de Extração, Indexação e Busca em Arquivos UFDR
 """
 
-__version__ = "1.0.0"
+__version__ = "1.1.3"
 __author__ = "Matheus C. Pestana"